4.23

+++ conrtollers/blogs.js
@@ -2,8 +2,8 @@ const blogsRouter = require('express').Router()
 const Blog = require('../moduls/blog')
 
 blogsRouter.get('/', async (request, response) => {
-  const result = await Blog.find({})
-    response.json(result)
+  const result = await Blog.find({}).populate('user', {username: 1, name: 1})
+  response.json(result)
 })
 
 blogsRouter.get('/:id', async (request, response) => {
@@ -16,26 +16,45 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    user.blogs = user.blogs.filter(a => a !== blog.id)
+    await user.save()
+
+    return response.status(204).end()
+  }
+  response
+    .status(401)
+    .json({ error: 'invalid user' })
 })
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
+  if(!request.token) {
+    response.status(401).json({ error: 'token missing' })
+  }
+  if(!request.user) {
+    response.status(401).json({ error: 'token invalid'})
+  }
 
-  if((body.title && body.url)) {
+
+  if (body.title && body.url) {
     const blog = new Blog({
       title: body.title,
       author: body.author,
+      user: user.id,
       url: body.url,
       likes: body.likes || 0
     })
   
-    blog.save()
-      .then(result => {
-        response.status(201).json(result)
-      })
-      
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
+    response.status(201).json(savedBlog)  
   } else {
     response.status(400).end()
   }
